Reverting conversion of AlternateTypeRule to an interface for simplcity
