+++ springfox-core/src/main/java/springfox/documentation/schema/AlternateTypeRule.java
@@ -16,12 +16,41 @@
  *
  *
  */
+
 package springfox.documentation.schema;
 
 import com.fasterxml.classmate.ResolvedType;
 import org.springframework.core.Ordered;
 
-public interface AlternateTypeRule extends Ordered {
+import static springfox.documentation.schema.WildcardType.*;
+
+public class AlternateTypeRule implements Ordered {
+  protected final ResolvedType original;
+  protected final ResolvedType alternate;
+  protected final int order;
+
+  /**
+   * Instantiates a new Alternate type rule.
+   *
+   * @param original  the original type
+   * @param alternate the alternate type
+   */
+  public AlternateTypeRule(ResolvedType original, ResolvedType alternate) {
+    this(original, alternate, Ordered.HIGHEST_PRECEDENCE);
+  }
+
+  /**
+   * Instantiates a new Alternate type rule.
+   *
+   * @param original  the original type
+   * @param alternate the alternate type
+   * @param order the order {@link Ordered} in which the rules are applied
+   */
+  public AlternateTypeRule(ResolvedType original, ResolvedType alternate, int order) {
+    this.original = original;
+    this.alternate = alternate;
+    this.order = order;
+  }
 
   /**
    * Provides alternate for supplier type.
@@ -29,7 +58,16 @@ public interface AlternateTypeRule extends Ordered {
    * @param type the type
    * @return the alternate for the type
    */
-  ResolvedType alternateFor(ResolvedType type);
+  public ResolvedType alternateFor(ResolvedType type) {
+    if (appliesTo(type)) {
+      if (hasWildcards(original)) {
+        return replaceWildcardsFrom(WildcardType.collectReplaceables(type, original), alternate);
+      } else {
+        return alternate;
+      }
+    }
+    return type;
+  }
 
   /**
    * Check if an alternate applies to type.
@@ -37,13 +75,17 @@ public interface AlternateTypeRule extends Ordered {
    * @param type the source
    * @return the boolean
    */
-  boolean appliesTo(ResolvedType type);
+  public boolean appliesTo(ResolvedType type) {
+    return hasWildcards(original)
+            && wildcardMatch(type, original)
+            || exactMatch(original, type);
+  }
 
-  /**
-   * Provides the type to use for the properties
-   *
-   * @return the type for retrieving properties
-   */
-  ResolvedType propertiesHost();
+  public ResolvedType propertiesHost() {
+    return alternate;
+  }
 
+  public int getOrder() {
+    return order;
+  }
 }
