+++ springfox-core/src/main/java/springfox/documentation/schema/RichAlternateTypeRule.java
@@ -22,7 +22,7 @@ package springfox.documentation.schema;
 import com.fasterxml.classmate.ResolvedType;
 import org.springframework.core.Ordered;
 
-public class RichAlternateTypeRule extends SimpleAlternateTypeRule {
+public class RichAlternateTypeRule extends AlternateTypeRule {
 
   public RichAlternateTypeRule(ResolvedType original, ResolvedType alternate) {
     this(original, alternate, Ordered.HIGHEST_PRECEDENCE);
